requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
markdown2==2.5.0
numpy==2.0.1
openai==1.41.0
pandas==2.2.2
praw==7.7.1
prawcore==2.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.2
websocket-client==1.8.0
